feature/#1006 Support couchbase collections
Current version is only supporting default-collection. With this development,
CouchbaseLockProvider is supporting different Collection with-in a Bucket and
this is provided by a class which is implemented by using
CouchbaseLockConfiguration interface.

Couchbase needed a new version because of Collections support. And which is why
it's updated from 3.0.5 to 3.2.0 and has not been observed any backward
compatibility issue.

README.md updated for these changes.